aws_instance: allow providing user data as base64
Providing base64 directly is convenient when it must be set to something
that isn't valid UTF-8, such as the gzipped payloads often passed to
cloud-init for more complex setups.

Passing binary data directly via Terraform attributes is not safe because
only UTF-8 strings can be stored in the state file. This is therefore
intended for use in conjunction with the base64 encoding mode of
template_cloudinit_config to enable gzip data to be passed safely to the
EC2 data without corruption as it passes through Terraform.