r/aws_instance: Fix associate_public_ip_address
Previously we had no way of determining if a boolean attribute was a
literal `false` or a zero-value type of `false`. With the core enhancement
added to Terraform via https://github.com/hashicorp/terraform/pull/15723,
we can now reliably check for a zero-type value of a boolean attribute,
or a literal boolean `false` inside of schema.

This allows the `associate_public_ip_address` to work as expected,
when configuring an AWS Instance inside of a subnet that defaults to
public address assignment, with a literal `false` for the `associate_public_ip_address` attribute.

```
$ make testacc TEST=./aws TESTARGS="-run=TestAccAWSInstance_explicitAssociatePublicAddress"
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go test ./aws -v -run=TestAccAWSInstance_explicitAssociatePublicAddress -timeout 120m
=== RUN   TestAccAWSInstance_explicitAssociatePublicAddress
--- PASS: TestAccAWSInstance_explicitAssociatePublicAddress (113.04s)
PASS
ok      github.com/terraform-providers/terraform-provider-aws/aws       113.045s
```